requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
py-cord==2.6.1
pymongo==4.11.2
python-dotenv==1.0.1
redis==5.2.1
yarl==1.18.3
